Modificando le modelo, cantida como decimal

+++ Models/IngresoEgresosCajaViewModel.cs
@@ -13,7 +13,7 @@ namespace mvcIpsa.Models
     public class IngresoEgresosCajaDetalleViewModel {
         public string cta_cuenta{ get; set; }
         public decimal precio { get; set; }
-        public int cantidad { get; set; }     
+        public decimal cantidad { get; set; }
         public decimal montodolar { get; set; }
     }
 }
